derive func type
